+++ verik-compiler/src/main/kotlin/io/verik/compiler/core/declaration/vk/CoreVkCoverCross.kt
@@ -18,13 +18,11 @@ package io.verik.compiler.core.declaration.vk
 
 import io.verik.compiler.ast.element.expression.common.ECallExpression
 import io.verik.compiler.ast.element.expression.common.EExpression
-import io.verik.compiler.ast.element.expression.common.ENothingExpression
 import io.verik.compiler.core.common.Core
 import io.verik.compiler.core.common.CoreScope
 import io.verik.compiler.core.common.TransformableCoreFunctionDeclaration
-import io.verik.compiler.message.Messages
 
-object CoreVkCoverPoint : CoreScope(Core.Vk.C_CoverPoint) {
+object CoreVkCoverCross : CoreScope(Core.Vk.C_CoverCross) {
 
     val F_bin_String_String = object : TransformableCoreFunctionDeclaration(
         parent,
@@ -33,8 +31,7 @@ object CoreVkCoverPoint : CoreScope(Core.Vk.C_CoverPoint) {
     ) {
 
         override fun transform(callExpression: ECallExpression): EExpression {
-            Messages.EXPRESSION_OUT_OF_CONTEXT.on(callExpression, name)
-            return ENothingExpression(callExpression.location)
+            return CoreVkCoverPoint.F_bin_String_String.transform(callExpression)
         }
     }