+++ verik-compiler/src/main/kotlin/io/verik/compiler/core/declaration/vk/CoreVkSpecial.kt
@@ -249,16 +249,25 @@ object CoreVkSpecial : CoreScope(CorePackage.VK) {
     val F_cp_Any_Function = object : TransformableCoreFunctionDeclaration(parent, "cp", "fun cp(Any, Function)") {
 
         override fun transform(callExpression: ECallExpression): EExpression {
-            Messages.EXPRESSION_OUT_OF_CONTEXT.on(callExpression, name)
-            return ENothingExpression(callExpression.location)
+            return F_cp_Any.transform(callExpression)
         }
     }
 
     val F_cc_Any = object : TransformableCoreFunctionDeclaration(parent, "cc", "fun cc(vararg CoverPoint)") {
 
         override fun transform(callExpression: ECallExpression): EExpression {
-            Messages.EXPRESSION_OUT_OF_CONTEXT.on(callExpression, name)
-            return ENothingExpression(callExpression.location)
+            return F_cp_Any.transform(callExpression)
+        }
+    }
+
+    val F_cc_Any_Function = object : TransformableCoreFunctionDeclaration(
+        parent,
+        "cc",
+        "fun cc(vararg CoverPoint, Function)"
+    ) {
+
+        override fun transform(callExpression: ECallExpression): EExpression {
+            return F_cp_Any.transform(callExpression)
         }
     }