+++ verik-core/src/main/kotlin/io/verik/core/CoverCross.kt
@@ -40,7 +40,7 @@ abstract class CoverCross {
     /**
      * Ignore a bin. [value] will be injected as SystemVerilog and used to specify the bin.
      */
-    fun ignoreBin(value: String) {
+    fun ignoreBin(name: String, value: String) {
         throw VerikException()
     }
 }