+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `Double` data type.
 - Support `monitor` and `strobe` functions.
 - Support `CoverGroup`, `CoverPoint`, and `CoverCross`.
+- Support cover bins for `CoverPoint` and `CoverCross`.
 
 ## [0.1.14]
 ### Importer